switch epoch-time to i64, wip conversation-list modal

+++ lumni/src/utils/time_parse_ext.rs
@@ -1,13 +1,13 @@
 use time::format_description::well_known::Rfc3339;
 use time::OffsetDateTime;
 
-pub fn rfc3339_to_epoch(timestamp: &str) -> Result<u64, time::Error> {
+pub fn rfc3339_to_epoch(timestamp: &str) -> Result<i64, time::Error> {
     let datetime = OffsetDateTime::parse(timestamp, &Rfc3339)?;
-    Ok(datetime.unix_timestamp() as u64)
+    Ok(datetime.unix_timestamp())
 }
 
-pub fn epoch_to_rfc3339(timestamp: u64) -> Result<String, time::Error> {
-    let datetime = OffsetDateTime::from_unix_timestamp(timestamp as i64)?;
+pub fn epoch_to_rfc3339(timestamp: i64) -> Result<String, time::Error> {
+    let datetime = OffsetDateTime::from_unix_timestamp(timestamp)?;
     Ok(datetime.to_string())
 }
